+++ src/Button/Button.stories.tsx
@@ -7,16 +7,47 @@ export default {
   component: Btn,
 } as Meta;
 
-export const Button: Story<ButtonProps> = (args) => (
-  <div className="section base-margin-top dbl-margin-bottom">
-    <h3 className="display-5">Button</h3>
-    <Btn {...args}>
-      {args.icon ? <span className="icon-check" /> : "Button"}
-    </Btn>
-  </div>
-);
+export const Button: Story<ButtonProps> = (args) => {
+  const [ref, setRef] = React.useState(null);
+
+  React.useEffect(() => {
+    console.log("Got the refs");
+    console.log({ ref });
+  }, [ref]);
+
+  return (
+    <div className="section base-margin-top dbl-margin-bottom">
+      <h3 className="display-5">Button</h3>
+      <Btn {...args} ref={setRef}>
+        {args.icon ? <span className="icon-check" /> : "Button"}
+      </Btn>
+    </div>
+  );
+};
 
 Button.args = {
   color: "primary",
   size: "default",
+};
+
+Button.argTypes = {
+  color: {
+    options: [
+      "primary",
+      "secondary",
+      "success",
+      "dark",
+      "ghost",
+      "link",
+      "light",
+      "danger",
+    ],
+    type: { name: "string", required: false },
+    control: "select",
+  },
+  size: {
+    options: ["small", "default", "large"],
+    type: { name: "string", required: false },
+    control: "radio",
+  },
 };