+++ src/Button/Button.tsx
@@ -1,4 +1,9 @@
-import React, { FC } from "react";
+import React, {
+  createElement,
+  forwardRef,
+  ForwardRefExoticComponent,
+  HTMLProps,
+} from "react";
 
 import { appendClass as ac } from "src/utils";
 
@@ -14,7 +19,7 @@ export type ButtonColor =
   | "light"
   | "danger";
 
-export interface ButtonProps extends React.ButtonHTMLAttributes<any> {
+export type ButtonProps = {
   size?: "small" | "default" | "large";
   color?: ButtonColor;
   wide?: boolean;
@@ -26,54 +31,85 @@ export interface ButtonProps extends React.ButtonHTMLAttributes<any> {
   className?: string;
   icon?: boolean;
   type?: "submit" | "reset" | "button";
+} & Omit<HTMLProps<HTMLButtonElement>, "size">;
+
+type BFR = ForwardRefExoticComponent<ButtonProps>;
+
+interface ButtonType extends BFR {
+  Primary: BFR;
+  Secondary: BFR;
+  Success: BFR;
+  Dark: BFR;
+  Ghost: BFR;
+  Link: BFR;
+  Light: BFR;
+  Danger: BFR;
 }
 
-interface ButtonColors {
-  Primary: FC<ButtonProps>;
-  Secondary: FC<ButtonProps>;
-  Success: FC<ButtonProps>;
-  Dark: FC<ButtonProps>;
-  Ghost: FC<ButtonProps>;
-  Link: FC<ButtonProps>;
-  Light: FC<ButtonProps>;
-  Danger: FC<ButtonProps>;
-}
+let Button: ButtonType;
 
-export const Button: ButtonColors & FC<ButtonProps> = ({
-  size = "default",
-  color = "primary",
-  wide = false,
-  justified = false,
-  circle = false,
-  className = null,
-  asLink = false,
-  style = null,
-  selected = false,
-  type = null,
-  icon = false,
-  ...props
-}) =>
-  React.createElement(asLink ? "a" : "button", {
-    className: `btn${ac(size !== "default", `btn--${size}`)} btn--${color}${ac(
-      wide,
-      "btn--wide"
-    )}${ac(justified, "btn--justified")}${ac(icon, "btn--icon")}${ac(
-      circle,
-      "btn--circle"
-    )}${ac(selected, "selected")}${ac(className)}${ac(
-      asLink,
-      "flex-middle flex-center"
-    )}`,
-    style: { ...(style || {}), ...(asLink ? { display: "flex" } : {}) },
-    ...(asLink ? {} : { type: type || "button" }),
-    ...props,
-  });
+(Button as any) = forwardRef<
+  HTMLButtonElement | HTMLAnchorElement,
+  ButtonProps
+>(
+  (
+    {
+      size = "default",
+      color = "primary",
+      wide = false,
+      justified = false,
+      circle = false,
+      className = null,
+      asLink = false,
+      style = null,
+      selected = false,
+      type = null,
+      icon = false,
+      ...props
+    },
+    ref
+  ) =>
+    createElement(asLink ? "a" : "button", {
+      className: `btn${ac(
+        size !== "default",
+        `btn--${size}`
+      )} btn--${color}${ac(wide, "btn--wide")}${ac(
+        justified,
+        "btn--justified"
+      )}${ac(icon, "btn--icon")}${ac(circle, "btn--circle")}${ac(
+        selected,
+        "selected"
+      )}${ac(className)}${ac(asLink, "flex-middle flex-center")}`,
+      style: { ...(style || {}), ...(asLink ? { display: "flex" } : {}) },
+      ...(asLink ? {} : { type: type || "button" }),
+      ...props,
+      ref,
+    })
+);
 
-Button.Primary = (props) => <Button {...props} color="primary" />;
-Button.Secondary = (props) => <Button {...props} color="secondary" />;
-Button.Success = (props) => <Button {...props} color="success" />;
-Button.Dark = (props) => <Button {...props} color="dark" />;
-Button.Ghost = (props) => <Button {...props} color="ghost" />;
-Button.Link = (props) => <Button {...props} color="link" />;
-Button.Light = (props) => <Button {...props} color="light" />;
-Button.Danger = (props) => <Button {...props} color="danger" />;
+Button.Primary = forwardRef((props, ref) => (
+  <Button {...props} color="primary" ref={ref} />
+));
+Button.Secondary = forwardRef((props, ref) => (
+  <Button {...props} color="secondary" ref={ref} />
+));
+Button.Success = forwardRef((props, ref) => (
+  <Button {...props} color="success" ref={ref} />
+));
+Button.Dark = forwardRef((props, ref) => (
+  <Button {...props} color="dark" ref={ref} />
+));
+Button.Ghost = forwardRef((props, ref) => (
+  <Button {...props} color="ghost" ref={ref} />
+));
+Button.Link = forwardRef((props, ref) => (
+  <Button {...props} color="link" ref={ref} />
+));
+Button.Light = forwardRef((props, ref) => (
+  <Button {...props} color="light" ref={ref} />
+));
+Button.Danger = forwardRef((props, ref) => (
+  <Button {...props} color="danger" ref={ref} />
+));
+
+export { Button };