+++ src/ReactSelect/components.tsx
@@ -1,4 +1,4 @@
-import React, { createElement } from "react";
+import React from "react";
 import type {
   MultiValueGenericProps,
   MultiValueRemoveProps,