+++ src/index.ts
@@ -38,3 +38,4 @@ export * from "./VSeparator";
 export * from "./Slider";
 export * from "./utils/base16Theme";
 export * from "./Kbd";
+export * from "./Blockquote";